joystick: added `SDL_JoystickCompatName`
The new function `SDL_JoystickCompatName` returns the name of the joystick without any normalization applied.
It is similar to the name returned by `SDL_JoystickName` pre-2.0.12.

NOTE: the function is only implemented for the Linux joystick driver.

The function is useful for applications that assumed the name returned by `SDL_JoysticktName` is the same name exposed by the Linux event/udev APIs.